requirements.txt
pysqlcipher3==1.2.0
python-dotenv==1.0.1
